``` Improve dtype handling and error management in training
- Changed `dtype` from `bfloat16` to `float32` in `train_daggpt_tiny.py` for improved stability.
- Updated `dtype` from `float32` to `bfloat16` in `train_default.py` as preferable for specific hardware.
- Enhanced dtype fallback logic in `train.py` with a validation step for `bfloat16` on CUDA devices.
- Implemented a mechanism to track and limit consecutive errors, terminating training on critical issues or too many errors.
- Added additional test cases in `test_train.py` to ensure error detection and handling logic functions correctly.
- Introduced a test for critical error detection to ensure proper termination for unrecoverable errors.
```

+++ config/train_daggpt_tiny.py
@@ -36,7 +36,7 @@ min_lr = 1e-5
 dataset = "proofpile"
 
 backend = "nccl"
-dtype = "bfloat16"  # keep this, fine on A100/H100
+dtype = "float32"
 compile = False  # set False to avoid extra compilation overhead
 
 always_save_checkpoint = True
